test(catalog): verify large exact profile wrapping
